% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save_as_csv
This function reads the csv file}
\usage{
save_as_csv(Data, File)
}
\arguments{
\item{File}{}
}
\description{
Save_as_csv
This function reads the csv file
}
